Stream all docker command output

+++ arrstack.sh
@@ -78,14 +78,9 @@ _exec_cmd() {
   fi
 
   local rc
-  if [[ "${argv[0]}" == docker && ( "${argv[1]:-}" == compose || "${argv[1]:-}" == pull ) ]]; then
-    if [[ -e /dev/fd/3 ]]; then
-      "${argv[@]}" 2>&1 | tee /dev/fd/3
-      rc=${PIPESTATUS[0]}
-    else
-      "${argv[@]}"
-      rc=$?
-    fi
+  if [[ "${argv[0]}" == docker && -e /dev/fd/3 ]]; then
+    "${argv[@]}" 2>&1 | tee /dev/fd/3
+    rc=${PIPESTATUS[0]}
   else
     "${argv[@]}"
     rc=$?
@@ -94,7 +89,7 @@ _exec_cmd() {
   if (( warn_on_fail )) && (( rc != 0 )); then
     warn "Command failed ($rc): $(_stringify_cmd "${argv[@]}")"
   fi
-  return $rc
+  return "$rc"
 }
 
 run() { _exec_cmd 0 "$@"; }
